% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsrtm.R
\name{getsrtm}
\alias{getsrtm}
\title{Gets the SRTM90 for Vietnam.}
\source{
The CGIAR consortium (\url{http://srtm.csi.cgiar.org}).
}
\usage{
getsrtm()
}
\description{
\code{getsrtm} returns a RasterLayer object that points to a GeoTIFF file on disk containing the SRTM90 for Vietnam.
}
\examples{
# getting the data:
srtm <- getsrtm()
srtm
filename(srtm)
projection(srtm)

# Ploting the elevation of Vietnam
plot(srtm)

# Ploting the elevation of Vietnam without legend and without axes:
plot(srtm, axes = FALSE, legend = FALSE)

# Plotting first the administrative limits of the country and then the elevations:
plot(gadmVN::gadm(level = "country"))
plot(srtm, axes = FALSE, legend = FALSE, add = TRUE)

# Cropping elevation inside one polygon:
hanoi <- subset(gadmVN::gadm(resolution = "high"), province == "Ha Noi")
hanoi_ele <- crop(srtm, hanoi)
plot(hanoi_ele)
plot(hanoi, add = TRUE)

# Removing anything that outside the polygon:
themask <- rasterize(hanoi, hanoi_ele)
hanoi_ele <- mask(hanoi_ele, themask)
plot(hanoi_ele)
plot(hanoi, add = TRUE)

# Extracting the elevation in 1 point:
extract(srtm, matrix(c(105, 21), 1))

# Extracting the elevations inside the province of Hanoi:
hanoi <- subset(gadmVN::gadm(resolution = "high"), province == "Ha Noi")
elevations <- extract(srtm, hanoi)
avg_ele_hanoi <- mean(unlist(elevations))

# Extracting the elevations in a buffer of 30 km of radius around Da Lat:
ele_dalat <- extract(srtm, matrix(c(108.459636, 11.938812), 1),
                     buffer = 30000)
mean(unlist(ele_dalat))

}
