% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srtm90}
\alias{srtm90}
\title{SRTM90 elevation data for Vietnam}
\format{An object of class \code{RasterLayer} of dimension 17796 x 8790 x 1.}
\source{
SRTM90 data base from \url{http://srtm.csi.cgiar.org}.
}
\usage{
srtm90
}
\description{
The SRTM90 Digital Elevation Model from CGIAR.
}
\details{
\code{srtm90} is a RasterLayer object that contains the elevation data for
Vietnam from the SRTM90.
}
\examples{
## Plotting the elevations of Vietnam:
\dontrun{
srtm90 <- getsrtm()
plot(srtm90)
}
## Extracting the elevation in one location:
## raster::extract
}
\author{
Marc Choisy
}
\keyword{datasets}
